# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/import-files/{accountNumber}": {
      "post": {
        "operationId": "ImportFilesController_uploadMultipleImportFiles",
        "parameters": [
          {
            "name": "accountNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "documentType",
                  "files"
                ],
                "properties": {
                  "documentType": {
                    "$ref": "#/components/schemas/ImportDocumentType"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UploadedImportFileDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "No files, missing or invalid documentType, or an unsafe file name."
          },
          "404": {
            "description": "No import case (mbl) with this account number."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "import-files"
        ]
      },
      "get": {
        "operationId": "ImportFilesController_listImportFiles",
        "parameters": [
          {
            "name": "accountNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UploadedImportFileDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "accountNumber is not a positive integer."
          },
          "404": {
            "description": "No import case (mbl) with this account number."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "import-files"
        ]
      }
    },
    "/api/import-files/{accountNumber}/files/{fileId}": {
      "get": {
        "operationId": "ImportFilesController_getImportFile",
        "parameters": [
          {
            "name": "accountNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file contents, with Content-Type set to its MIME type.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "accountNumber or fileId is not a positive integer."
          },
          "404": {
            "description": "Unknown case, no such file in that case, or the file is missing on disk."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "import-files"
        ]
      }
    },
    "/api/import-files/{accountNumber}/line-items": {
      "get": {
        "operationId": "ImportFilesController_getSupplierInvoiceLineItems",
        "parameters": [
          {
            "name": "accountNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceLineItemDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "accountNumber is not a positive integer."
          },
          "404": {
            "description": "No import case (mbl) with this account number."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "import-files"
        ]
      }
    },
    "/api/import-files/{accountNumber}/line-items/classification": {
      "put": {
        "operationId": "ImportFilesController_saveLineItemClassifications",
        "parameters": [
          {
            "name": "accountNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveLineItemClassificationsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceLineItemDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload, a line index outside its file, or an unknown country id."
          },
          "404": {
            "description": "Unknown case, or a fileId that is not a file of that case."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "import-files"
        ]
      }
    },
    "/api/mbl": {
      "get": {
        "operationId": "MblController_findAll",
        "parameters": [
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "integer"
            }
          },
          {
            "name": "carrierName",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Maersk",
              "type": "string"
            }
          },
          {
            "name": "caseNumber",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MblSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "mbl"
        ]
      },
      "post": {
        "operationId": "MblController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMblDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MblDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "mbl"
        ]
      }
    },
    "/api/mbl/{id}": {
      "get": {
        "operationId": "MblController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MblDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "mbl"
        ]
      },
      "patch": {
        "operationId": "MblController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMblDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MblDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "mbl"
        ]
      },
      "delete": {
        "operationId": "MblController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "mbl"
        ]
      }
    },
...
  },
  "info": {
    "title": "Amilut API",
    "description": "Customs brokerage back office API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ImportDocumentType": {
        "type": "string",
        "enum": [
          "SUPPLIER_INVOICE",
          "BILL_OF_LADING",
          "MASTER_BILL_OF_LADING",
          "PACKING_LIST",
          "CERTIFICATE_OF_ORIGIN"
        ],
        "description": "Kind of paperwork this file represents.",
        "x-enum-varnames": [
          "SUPPLIER_INVOICE",
          "BILL_OF_LADING",
          "MASTER_BILL_OF_LADING",
          "PACKING_LIST",
          "CERTIFICATE_OF_ORIGIN"
        ]
      },
...
      "LineItemClassificationDto": {
        "type": "object",
        "properties": {
          "tradeAgreement": {
            "nullable": true,
            "example": "eu",
            "allOf": [
              {
                "$ref": "#/components/schemas/TradeAgreement"
              }
            ]
          },
          "classificationCode": {
            "type": "string",
            "example": "8539.50.00",
            "maxLength": 100
          },
          "approvals": {
            "type": "array",
            "example": [
              "standard_or_declaration"
            ],
            "items": {
              "$ref": "#/components/schemas/ClassificationApproval"
            }
          },
          "licenses": {
            "type": "array",
            "example": [
              "health_pharmacy_division"
            ],
            "items": {
              "$ref": "#/components/schemas/ClassificationLicense"
            }
          },
          "countryId": {
            "type": "integer",
            "nullable": true,
            "example": 106,
            "description": "`countries.id` of the origin country."
          }
        },
        "required": [
          "tradeAgreement",
          "classificationCode",
          "approvals",
          "licenses",
          "countryId"
        ]
      },
      "InvoiceLineItemDto": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "integer",
            "example": 42,
            "description": "The `import_account_files` row (supplier invoice) this line came from."
          },
          "lineIndex": {
            "type": "integer",
            "example": 0,
            "description": "Position of the line inside that file's `lineItems`."
          },
          "item": {
            "type": "string",
            "example": "Y8022-140BK",
            "description": "Item / part / model code as printed (may contain spaces)."
          },
          "description": {
            "type": "string",
            "example": "Light Fixtures"
          },
          "quantity": {
            "type": "number",
            "nullable": true,
            "example": 15
          },
          "price": {
            "type": "number",
            "nullable": true,
            "example": 15.32,
            "description": "Unit price with the currency mark stripped."
          },
          "total": {
            "type": "number",
            "nullable": true,
            "example": 229.8,
            "description": "Line total with the currency mark stripped."
          },
          "classification": {
            "description": "What the classification screen recorded; empty defaults when never classified.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LineItemClassificationDto"
              }
            ]
          }
        },
        "required": [
          "fileId",
          "lineIndex",
          "item",
          "description",
          "quantity",
          "price",
          "total",
          "classification"
        ]
      },
      "UploadedImportFileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 42,
            "description": "Id of the `import_account_files` row created for this file."
          },
          "documentType": {
            "description": "Kind of paperwork this file represents.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ImportDocumentType"
              }
            ]
          },
          "name": {
            "type": "string",
            "example": "חשבון ספק.pdf",
            "description": "Original file name, as uploaded."
          },
          "size": {
            "type": "integer",
            "example": 870400,
            "description": "Size in bytes."
          },
          "mimeType": {
            "type": "string",
            "example": "application/pdf"
          },
          "relativePath": {
            "type": "string",
            "example": "1000/חשבון ספק.pdf",
            "description": "Path relative to the storage root (`storage/` at the repo root)."
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-22T16:05:18.000Z"
          },
          "lineItems": {
            "description": "Goods lines extracted from the file. Only present for SUPPLIER_INVOICE; empty when nothing could be extracted (see `extractionError`).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineItemDto"
            }
          },
          "extractionError": {
            "type": "string",
            "example": "No line-item table found",
            "description": "Why `lineItems` is empty: unreadable file, unsupported type or no table. Absent when extraction succeeded or was not attempted."
          }
        },
        "required": [
          "id",
          "documentType",
          "name",
          "size",
          "mimeType",
          "relativePath",
          "uploadedAt"
        ]
      },
...
      "SaveLineItemClassificationsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItemClassificationUpdateDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "MblShippingType": {
        "type": "string",
        "enum": [
          "sea",
          "air"
        ],
        "x-enum-varnames": [
          "SEA",
          "AIR"
        ]
      },
      "SeaMethod": {
        "type": "string",
        "enum": [
          "fcl_fcl",
          "fcl_lcl",
          "lcl_lcl",
          "groupage_fcl"
        ],
        "x-enum-varnames": [
          "FCL_FCL",
          "FCL_LCL",
          "LCL_LCL",
          "GROUPAGE_FCL"
        ]
      },
      "MblSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "shippingType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MblShippingType"
              }
            ]
          },
          "seaMethod": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SeaMethod"
              }
            ]
          },
          "mblNumber": {
            "type": "string",
            "nullable": true,
            "example": "MBL-987654321"
          },
          "carrierName": {
            "type": "string",
            "nullable": true,
            "example": "MAERSK LINE"
          },
          "hblCount": {
            "type": "integer",
            "example": 3
          },
          "orderIds": {
            "example": [
              1000,
              1002
            ],
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "customerNames": {
            "example": [
              "ACME Ltd."
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-01T08:30:00.000Z"
          }
        },
        "required": [
          "id",
          "shippingType",
          "seaMethod",
          "mblNumber",
          "carrierName",
          "hblCount",
          "orderIds",
          "customerNames",
          "createdAt"
        ]
      },
...
      "CreateMblDto": {
        "type": "object",
        "properties": {
          "shippingType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MblShippingType"
              }
            ]
          },
          "seaMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SeaMethod"
              }
            ]
          },
          "customerId": {
            "type": "integer",
            "example": 3
          },
          "mblNumber": {
            "type": "string",
            "example": "MBL-987654321"
          },
          "bookingNumber": {
            "type": "string",
            "example": "BKG-456789"
          },
          "vesselName": {
            "type": "string",
            "example": "EVER GLORY"
          },
          "voyageNumber": {
            "type": "string",
            "example": "062W"
          },
          "portOfLoading": {
            "type": "string",
            "example": "Shanghai, China"
          },
          "portOfDischarge": {
            "type": "string",
            "example": "Haifa, Israel"
          },
          "finalDestination": {
            "type": "string",
            "example": "Haifa, Israel"
          },
          "shipperName": {
            "type": "string",
            "example": "GLOBAL CONSOLIDATION LTD."
          },
          "shipperAddress": {
            "type": "string",
            "example": "No. 12, Seaport Road, Shanghai, China"
          },
          "consigneeName": {
            "type": "string",
            "example": "TO ORDER OF MAERSK LINE (As per House B/L)"
          },
          "consigneeAddress": {
            "type": "string"
          },
          "notifyPartyName": {
            "type": "string",
            "example": "EREZ LOGISTICS LTD."
          },
          "notifyPartyAddress": {
            "type": "string",
            "example": "Haifa, Israel"
          },
          "containerNumber": {
            "type": "string",
            "example": "MSKU1234567"
          },
          "containerSealNumber": {
            "type": "string",
            "example": "TCLU1234567"
          },
          "cargoDescription": {
            "type": "string",
            "example": "CONSOLIDATED CARGO"
          },
          "grossWeightKg": {
            "type": "number",
            "example": 5400
          },
          "volumeCbm": {
            "type": "number",
            "example": 25.5
          },
          "freightTerms": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentTerms"
              }
            ]
          },
          "receiptDeliveryType": {
            "type": "string",
            "example": "CY/CFS"
          },
          "placeOfIssue": {
            "type": "string",
            "example": "SHANGHAI"
          },
          "dateOfIssue": {
            "type": "string",
            "format": "date",
            "example": "2026-09-20"
          },
          "carrierName": {
            "type": "string",
            "example": "MAERSK LINE"
          },
          "containers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMblContainerDto"
            }
          }
        },
        "required": [
          "shippingType"
        ]
      },
...
      "UpdateMblDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "example": 3
          },
          "mblNumber": {
            "type": "string",
            "example": "MBL-987654321"
          },
          "bookingNumber": {
            "type": "string",
            "example": "BKG-456789"
          },
          "vesselName": {
            "type": "string",
            "example": "EVER GLORY"
          },
          "voyageNumber": {
            "type": "string",
            "example": "062W"
          },
          "portOfLoading": {
            "type": "string",
            "example": "Shanghai, China"
          },
          "portOfDischarge": {
            "type": "string",
            "example": "Haifa, Israel"
          },
          "finalDestination": {
            "type": "string",
            "example": "Haifa, Israel"
          },
          "shipperName": {
            "type": "string",
            "example": "GLOBAL CONSOLIDATION LTD."
          },
          "shipperAddress": {
            "type": "string",
            "example": "No. 12, Seaport Road, Shanghai, China"
          },
          "consigneeName": {
            "type": "string",
            "example": "TO ORDER OF MAERSK LINE (As per House B/L)"
          },
          "consigneeAddress": {
            "type": "string"
          },
          "notifyPartyName": {
            "type": "string",
            "example": "EREZ LOGISTICS LTD."
          },
          "notifyPartyAddress": {
            "type": "string",
            "example": "Haifa, Israel"
          },
          "containerNumber": {
            "type": "string",
            "example": "MSKU1234567"
          },
          "containerSealNumber": {
            "type": "string",
            "example": "TCLU1234567"
          },
          "cargoDescription": {
            "type": "string",
            "example": "CONSOLIDATED CARGO"
          },
          "grossWeightKg": {
            "type": "number",
            "example": 5400
          },
          "volumeCbm": {
            "type": "number",
            "example": 25.5
          },
          "freightTerms": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentTerms"
              }
            ]
          },
          "receiptDeliveryType": {
            "type": "string",
            "example": "CY/CFS"
          },
          "placeOfIssue": {
            "type": "string",
            "example": "SHANGHAI"
          },
          "dateOfIssue": {
            "type": "string",
            "format": "date",
            "example": "2026-09-20"
          },
          "carrierName": {
            "type": "string",
            "example": "MAERSK LINE"
          },
          "containers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMblContainerDto"
            }
          }
        }
      },
...
    }
  }
}